refactor: extracted Habits subviews to separate files, NameField and IconPicker to Components
