+++ Nudge/Views/Habits/HabitsView.swift
@@ -71,46 +71,3 @@ struct HabitsView: View {
         }
     }
 }
-
-//==== HabitsHeader =============================================
-
-private struct HabitsHeader: View {
-
-    var body: some View {
-        HStack {
-            VStack(alignment: .leading, spacing: Spacing.xs) {
-                Text(String(localized: "habits_title"))
-                    .font(.system(size: FontSize.display, weight: .bold))
-                    .foregroundStyle(Theme.nudgeTextPrimary)
-                Text(String(localized: "habits_subtitle"))
-                    .font(.system(size: FontSize.sm))
-                    .foregroundStyle(Theme.nudgeTextMuted)
-            }
-            Spacer()
-        }
-    }
-}
-
-//==== AddHabitButton =============================================
-
-private struct AddHabitButton: View {
-
-    let action: () -> Void
-
-    var body: some View {
-        Button(action: action) {
-            HStack(spacing: Spacing.sm) {
-                Image(systemName: "plus.circle")
-                    .font(.system(size: IconSize.lg))
-                    .foregroundStyle(Theme.nudgeAccentLight)
-                Text(String(localized: "habits_add_placeholder"))
-                    .font(.system(size: FontSize.lg))
-                    .foregroundStyle(Theme.nudgeTextMuted)
-                Spacer()
-            }
-            .padding(Spacing.lg)
-            .background(Theme.nudgeSurface)
-            .clipShape(RoundedRectangle(cornerRadius: Radius.lg))
-        }
-    }
-}
